+++ backend/Dtos/SavingGoal/UpdateSavingGoalDto.cs
@@ -3,6 +3,6 @@ namespace api.Dtos.SavingGoal
   public class UpdateSavingGoalDto
   {
     public String Name { get; set; } = String.Empty;
-    public int TargetAmount { get; set; }
+    public decimal TargetAmount { get; set; }
   }
 }
