+++ backend/Models/SavingGoal.cs
@@ -12,8 +12,8 @@ namespace api.Models
     public User? User { get; set; }
 
 
-    public int TargetAmount { get; set; }
-    public int CurrentSaved { get; set; }
+    public decimal TargetAmount { get; set; }
+    public decimal CurrentSaved { get; set; }
     public DateTime CreatedAt { get; set; } = DateTime.Now;
   }
 }
